+++ SurfaceTopography/IO/Reader.py
@@ -26,19 +26,17 @@
 #
 
 import abc
-import warnings
 
 import numpy as np
 
 
 class ChannelInfo:
     """
-    Information on topography channels contained within a file. The interface
-    is identical to :obj:`HeightContainer` and subclasses.
+    Information on topography channels contained within a file.
     """
 
     def __init__(self, reader, index, name=None, dim=None, nb_grid_pts=None,
-                 physical_sizes=None, periodic=None,
+                 physical_sizes=None, height_scale_factor=None, periodic=None,
                  info={}):
         """
         Initialize the channel. Use as many information from the file as
@@ -59,8 +57,10 @@ class ChannelInfo:
             Number grid points in each dimension.
         physical_sizes: tuple of floats
             Physical dimensions.
+        height_scale_factor: float
+            Number by which all heights have been multiplied.
         periodic: bool
-            Wether the SurfaceTopography should be interpreted as one period of
+            Whether the SurfaceTopography should be interpreted as one period of
             a periodic surface. This will affect the PSD and autocorrelation
             calculations (windowing).
         info: dict
@@ -75,6 +75,7 @@ class ChannelInfo:
             if nb_grid_pts is None else tuple(np.ravel(nb_grid_pts))
         self._physical_sizes = None \
             if physical_sizes is None else tuple(np.ravel(physical_sizes))
+        self._height_scale_factor = height_scale_factor
         self._periodic = periodic
         self._info = info.copy()
 
@@ -95,12 +96,14 @@ class ChannelInfo:
             physical size, then this parameter will override the physical
             size found in the data file.
         height_scale_factor : float
-            Override height scale factor found in the data file.
+            Factor by which the heights should be multiplied.
+            This parameter is only available for file formats that do not provide
+            metadata information about units and associated length scales.
         info : dict
             This dictionary will be appended to the info dictionary returned
             by the reader.
         periodic: bool
-            Wether the SurfaceTopography should be interpreted as one period of
+            Whether the SurfaceTopography should be interpreted as one period of
             a periodic surface. This will affect the PSD and autocorrelation
             calculations (windowing)
         subdomain_locations : tuple of ints
@@ -131,10 +134,10 @@ class ChannelInfo:
 
     @property
     def name(self):
-        '''
+        """
         Name of the channel.
         Can be used in a UI for identifying a channel.
-        '''
+        """
         return self._name
 
     @property
@@ -167,6 +170,21 @@ class ChannelInfo:
         `topography` method that returns the topography object.
         """
         return self._physical_sizes
+
+    @property
+    def height_scale_factor(self):
+        """
+        If a height scale factor can be determined from the file, a float is
+        returned.
+
+        If no height scale factor can be determined from the file, then
+        None is returned.
+
+        Note that the height scale factor obtained from the file can be
+        overwritten by passing a `height_scale_factor` argument to the
+        `topography` method that returns the topography object.
+        """
+        return self._height_scale_factor
 
     @property
     def is_periodic(self):
@@ -203,7 +221,7 @@ class ChannelInfo:
         """
         The area per point is returned as the product over the pixel size
         tuple. If `pixel_size` returns None, than also `area_per_pt` returns
-        None.s
+        None.
         """
         if self.pixel_size is None:
             return None
@@ -295,25 +313,29 @@ class ReaderBase(metaclass=abc.ABCMeta):
         return self.channels[self._default_channel_index]
 
     @classmethod
-    def _check_physical_sizes(self, physical_sizes_from_arg,
+    def _check_physical_sizes(cls, physical_sizes_from_arg,
                               physical_sizes=None):
         """Handle overriding of `physical_sizes` arguments and make sure,
-        that return value is not `None`."""
+        that return value is not `None`.
+
+        Parameters
+        ----------
+        physical_sizes_from_arg: tuple or float or None
+            physical sizes given as argument when instantiating a topography from a reader
+        physical_sizes: tuple or float or None
+            physical sizes which were present in the file
+        """
         if physical_sizes is None:
             if physical_sizes_from_arg is None:
                 raise ValueError("`physical_sizes` could not be extracted from file, you must provide it.")
+            eff_physical_sizes = physical_sizes_from_arg
+        elif physical_sizes_from_arg is not None:
+            # in general this should result in an exception now
+            raise MetadataAlreadyFixedByFile('physical_sizes')
         else:
-            if physical_sizes_from_arg is None:
-                physical_sizes_from_arg = physical_sizes
-            if tuple(physical_sizes_from_arg) != tuple(physical_sizes):
-                warnings.warn(
-                    "A physical size different from the value specified when "
-                    "calling the reader was present in the file. We will "
-                    "ignore the value given in the file. Specified values: "
-                    "{}; Values from file: {}".format(
-                        physical_sizes,
-                        physical_sizes_from_arg))
-        return physical_sizes_from_arg
+            eff_physical_sizes = physical_sizes
+
+        return eff_physical_sizes
 
     @abc.abstractmethod
     def topography(self, channel_index=None, physical_sizes=None,
@@ -323,14 +345,18 @@ class ReaderBase(metaclass=abc.ABCMeta):
         """
         Returns an instance of a subclass of :obj:`HeightContainer` that
         contains the topography data. The method allows to override data
-        found in the data file.
+        found in the data file by giving appropriate arguments.
+
+        For some of the given arguments it is checked whether they
+        have already been defined in the file, e.g. the height
+        scale factor, and if yes, an exception is raised (see below).
 
         Arguments
         ---------
         channel_index : int
             Index of the channel to load. See also `channels` method.
             (Default: None, which load the default channel)
-        physical_si""zes : tuple of floats
+        physical_sizes : tuple of floats
             Physical size of the topography. It is necessary to specify this
             if no physical size is found in the data file. If there is a
             physical size, then this parameter will override the physical
@@ -343,7 +369,7 @@ class ReaderBase(metaclass=abc.ABCMeta):
         periodic: bool
             Whether the SurfaceTopography should be interpreted as one period of
             a periodic surface. This will affect the PSD and autocorrelation
-            calculations (windowing)
+            calculations (windowing).
         subdomain_locations : tuple of ints
             Origin (location) of the subdomain handled by the present MPI
             process.
@@ -355,6 +381,12 @@ class ReaderBase(metaclass=abc.ABCMeta):
         -------
         topography : subclass of :obj:`HeightContainer`
             The object containing the actual topography data.
+
+        Raises
+        ------
+        MetadataAlreadyDefined
+            Raised if `physical_sizes` or `height_scale_factor` have already
+            been defined in the file, because they should not be overridden.
         """
         raise NotImplementedError
 
@@ -388,3 +420,29 @@ class CorruptFile(ReadFileError):
     but there is a mistake, for example the number of points doesn't match
     """
     pass
+
+
+class MetadataAlreadyFixedByFile(ReadFileError):
+    """
+    Raised when instantiating a topography from a reader,
+    given metadata which cannot be overridden, because
+    it is already fixed by the file contents.
+    """
+    def __init__(self, kw, alt_msg=None):
+        """
+        Parameters
+        ----------
+        kw: str
+            Name of the keyword argument to .topography().
+        alt_msg: str or None
+            If not None, use this as error message instead of
+            the default one.
+        """
+        self._kw = kw
+        self._alt_msg = alt_msg
+
+    def __str__(self):
+        if self._alt_msg:
+            return self._alt_msg
+        else:
+            return f"Value for keyword '{self._kw}' is already fixed by file contents and cannot be overridden"