+++ SurfaceTopography/ChangeLog.md
@@ -2,14 +2,18 @@
 Change log for SurfaceTopography
 =============================
 
-v0.94.0 (not published yet)
+v0.94.0 (10Jun21)
 ---------------------------
 
 - ENH: Added API tests for checking existence of functions
   and properties for different kinds of topographies (#100)
-- BUG: Fixes missing function for RMS curvature for uniform
+- ENH: Added height_scale_factor to reader channels (#98)
+- ENH: Disallow to provide physical_sizes or height_scale_factor
+  as argument to .topography() if fixed by file contents  
+- BUG: Fixed missing function for RMS curvature for uniform
   line scans (#95)
-- BUG: Fixes loading of 2D measurements from xyz data (#93)
+- BUG: Make sure that .topography(height_scale_factor=..) is effective  
+- BUG: Fixed loading of 2D measurements from xyz data (#93)
 
 v0.93.0 (27Apr21)
 -----------------